python/week04/backend/order_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
aio-pika